Finish defining get_slice_data in crud
