+++ breadbox/breadbox/crud/dataset.py
@@ -32,7 +32,7 @@ from ..schemas.custom_http_exception import (
     UserError,
 )
 from breadbox.crud.access_control import user_has_access_to_group
-from ..models.dataset import (
+from breadbox.models.dataset import (
     AnnotationType,
     Dataset,
     MatrixDataset,
@@ -54,10 +54,17 @@ from breadbox.crud.group import (
     TRANSIENT_GROUP_ID,
     get_transient_group,
 )
-from ..io.filestore_crud import get_slice, delete_data_files
+from breadbox.io.filestore_crud import (
+    get_slice,
+    delete_data_files,
+    get_feature_slice,
+    get_sample_slice,
+)
 from .metadata import cast_tabular_cell_value_type
 from .dataset_reference import add_id_mapping
 import typing
+
+from depmap_compute.slice import SliceQuery
 
 log = logging.getLogger(__name__)
 
@@ -1731,3 +1738,72 @@ def get_subsetted_matrix_dataset_df(
         df = df.rename(index=label_by_id)
 
     return df
+
+
+def get_slice_data(
+    db: SessionWithUser, filestore_location: str, slice_query: SliceQuery
+) -> pd.Series:
+    """
+    Loads data for the given slice query. 
+    The result will be a pandas series indexed by sample/feature ID 
+    (regardless of the identifier_type used in the query).
+    """
+    dataset_id = slice_query.dataset_id
+    dataset = get_dataset(db=db, user=db.user, dataset_id=dataset_id)
+    if dataset is None:
+        raise ResourceNotFoundError("Dataset not found")
+
+    if slice_query.identifier_type == "feature_id":
+        feature = get_dataset_feature_by_given_id(
+            db=db,
+            user=db.user,
+            dataset_id=dataset_id,
+            feature_label=slice_query.identifier,
+        )
+        slice_data = get_feature_slice(dataset, [feature.index], filestore_location)
+
+    elif slice_query.identifier_type == "feature_label":
+        feature = get_dataset_feature_by_label(
+            db=db,
+            user=db.user,
+            dataset_id=dataset_id,
+            feature_label=slice_query.identifier,
+        )
+        slice_data = get_feature_slice(dataset, [feature.index], filestore_location)
+
+    elif slice_query.identifier_type == "sample_id":
+        sample = get_dataset_sample_by_given_id(
+            db, db.user, dataset_id=dataset_id, sample_label=slice_query.identifier
+        )
+        slice_data = get_sample_slice(dataset, [sample.index], filestore_location)
+
+    elif slice_query.identifier_type == "sample_label":
+        sample = get_dataset_sample_by_label(
+            db, db.user, dataset_id=dataset_id, sample_label=slice_query.identifier
+        )
+        slice_data = get_sample_slice(dataset, [sample.index], filestore_location)
+
+    elif slice_query.identifier_type == "column":
+        if not dataset.format == "tabular_dataset":
+            raise UserError(
+                f"The slice query identifier type `column` may only be used with tabular datasets."
+            )
+        tabular_dimension_info = TabularDimensionsInfo(columns=[slice_query.identifier])
+        slice_data = get_subsetted_tabular_dataset_df(
+            db=db,
+            user=db.user,
+            dataset=dataset,
+            tabular_dimensions_info=tabular_dimension_info,
+            strict=True,
+        )
+
+    else:
+        raise ResourceNotFoundError(
+            f"Unrecognized slice query identifier type: `{slice_query.identifier_type}`",
+        )
+
+    if slice_data.empty or slice_data is None:
+        raise ResourceNotFoundError("No data matches the given slice query.")
+
+    # Convert the single-col/row DataFrame into a series and drop null values
+    return slice_data.dropna().squeeze()
